fix : showLanguage
